Extract shared state layout contracts
