+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "crates/celox-analysis",
     "crates/celox-design",
     "crates/celox-sir",
+    "crates/celox-state-layout",
     "crates/celox-bench-sv",
     "crates/celox-macros",
     "crates/celox-napi",
