+++ crates/celox-state-layout/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name                  = "celox-state-layout"
+version.workspace     = true
+authors.workspace     = true
+repository.workspace  = true
+license.workspace     = true
+description           = "Physical simulation-state layout contracts for Celox"
+edition.workspace     = true
+
+[lints]
+workspace = true
