+++ crates/celox/Cargo.toml
@@ -17,6 +17,7 @@ celox-macros          = { path = "../celox-macros" }
 celox-analysis        = { path = "../celox-analysis" }
 celox-design          = { path = "../celox-design" }
 celox-sir             = { path = "../celox-sir" }
+celox-state-layout    = { path = "../celox-state-layout" }
 fxhash                = { workspace = true }
 thiserror             = { workspace = true }
 clap                  = { workspace = true }
